Computer Vision/Conveyor Part Defect Detector/Source Code/requirements.txt
# Conveyor Part Defect Detector
ultralytics>=8.0
opencv-python>=4.8
numpy>=1.24
pyyaml>=6.0
